+++ src/Symfony/Component/Messenger/Bridge/Amqp/Transport/Connection.php
@@ -433,6 +433,27 @@ class Connection
         return null;
     }
 
+    /**
+     * Consume a message from the specified queue in blocking mode.
+     *
+     * @param ?callable(\AMQPEnvelope,\AMQPQueue):?false $callback If callback return false, then processing thread will be
+     * returned from AMQPQueue::consume() to PHP script. If null is passed, then the messages delivered to this client
+     * will be made available to the first real callback registered. That allows one to have a single callback
+     * consuming from multiple queues.
+     *
+     * @throws \AMQPException
+     */
+    public function pull(string $queueName, ?callable $callback): void
+    {
+        $this->clearWhenDisconnected();
+
+        if ($this->autoSetupExchange) {
+            $this->setupExchangeAndQueues();
+        }
+
+        $this->queue($queueName)->consume($callback);
+    }
+
     public function ack(\AMQPEnvelope $message, string $queueName): bool
     {
         return $this->queue($queueName)->ack($message->getDeliveryTag());